Keep cleanup and perf baselines clippy-clean
Consolidate the cleanup pass behind the repo's strict clippy gate, including perf matrix contract coverage, safer test helpers, wire compatibility fixes, and journal/perf documentation updates.

Constraint: cargo clippy --all-targets --features wire,test-hooks -- -D warnings must pass

Rejected: Add broad lint suppressions | avoidable warnings should be fixed at the source

Confidence: high

Scope-risk: broad

Directive: Preserve the strict clippy gate before future cleanup, perf, or wire changes

Tested: cargo clippy --all-targets --features wire,test-hooks -- -D warnings; cargo fmt --check; git diff --check

Not-tested: Full cargo test suite and Jepsen were not rerun for this commit

+++ src/wire/tests/protocol.rs
@@ -138,7 +138,7 @@ fn invalid_checksum_rejected() {
 #[test]
 fn oversized_message_rejected() {
     // Build a header claiming the message is 49 MiB.
-    let claimed_size = (49 * 1024 * 1024) as i32;
+    let claimed_size = 49 * 1024 * 1024;
     let header = MsgHeader {
         message_length: claimed_size,
         request_id: 1,
